refactor: rename Post component
