+++ components/Post/index.ts
@@ -1,0 +1,1 @@
+export { default as Post } from "./Post";
